Some updates―renaming interfaces and files
