+++ src/interactions/mod.ts
@@ -1,2 +1,2 @@
-export * from "./interactions.ts";
+export * from "./server.ts";
 export * from "./types/mod.ts";
