--- src/util/cdn.ts
@@ -1,10 +1,0 @@
-import { ImageFormats, ImageSize } from "../types/types.ts";
-
-export const formatImageURL = (
-  url: string,
-  size: ImageSize = 128,
-  format?: ImageFormats,
-) => {
-  return `${url}.${format ||
-    (url.includes("/a_") ? "gif" : "jpg")}?size=${size}`;
-};
