+++ src/util/utils.ts
@@ -1,6 +1,11 @@
-import { sendGatewayCommand } from "../ws/shard_manager.ts";
-import { ActivityType, StatusType } from "../types/types.ts";
 import { encode } from "../../deps.ts";
+import {
+  ActivityType,
+  ImageFormats,
+  ImageSize,
+  StatusType,
+} from "../types/types.ts";
+import { sendGatewayCommand } from "../ws/shard_manager.ts";
 
 export const sleep = (timeout: number) => {
   return new Promise((resolve) => setTimeout(resolve, timeout));
@@ -40,3 +45,12 @@ export function delay(ms: number): Promise<void> {
     }, ms)
   );
 }
+
+export const formatImageURL = (
+  url: string,
+  size: ImageSize = 128,
+  format?: ImageFormats,
+) => {
+  return `${url}.${format ||
+    (url.includes("/a_") ? "gif" : "jpg")}?size=${size}`;
+};
